RUMM-2401 Add tests to DatadogProviderCodepush
